Replace cosf/sinf with std::cos/std::sin in rotation matrices
RotationX, RotationY, and RotationZ used cosf/sinf which silently
truncates double-precision angles to float before computing trig
functions. Use std::cos/std::sin which dispatch to the correct
precision based on the template parameter T.

+++ include/mathfu/matrix.h
@@ -790,7 +790,7 @@ class Matrix {
   /// @param angle Angle (in radians).
   /// @return Matrix containing the result.
   static inline Matrix<T, 3> RotationX(T angle) {
-    return RotationX(Vector<T, 2>(cosf(angle), sinf(angle)));
+    return RotationX(Vector<T, 2>(std::cos(angle), std::sin(angle)));
   }
 
   /// @brief Create a 3x3 rotation Matrix from an angle (in radians) around
@@ -799,7 +799,7 @@ class Matrix {
   /// @param angle Angle (in radians).
   /// @return Matrix containing the result.
   static inline Matrix<T, 3> RotationY(T angle) {
-    return RotationY(Vector<T, 2>(cosf(angle), sinf(angle)));
+    return RotationY(Vector<T, 2>(std::cos(angle), std::sin(angle)));
   }
 
   /// @brief Create a 3x3 rotation Matrix from an angle (in radians)
@@ -808,7 +808,7 @@ class Matrix {
   /// @param angle Angle (in radians).
   /// @return Matrix containing the result.
   static inline Matrix<T, 3> RotationZ(T angle) {
-    return RotationZ(Vector<T, 2>(cosf(angle), sinf(angle)));
+    return RotationZ(Vector<T, 2>(std::cos(angle), std::sin(angle)));
   }
 
   /// @brief Create a 4x4 perspective Matrix.
